+++ tests/Predis/Command/Redis/TimeSeries/TSMRANGE_Test.php
@@ -305,7 +305,7 @@ class TSMRANGE_Test extends PredisCommandTestCase
      * @group connected
      * @group relay-resp3
      * @return void
-     * @requiresRedisVersion >= 8.10.0
+     * @requiresRedisVersion >= 8.9.0
      */
     public function testQueryRangeExcludesEmptySeriesWithExcludeEmptyModifier(): void
     {
@@ -348,7 +348,7 @@ class TSMRANGE_Test extends PredisCommandTestCase
      * @group connected
      * @group relay-resp3
      * @return void
-     * @requiresRedisVersion >= 8.10.0
+     * @requiresRedisVersion >= 8.9.0
      */
     public function testQueryRangeExcludesEmptySeriesWithExcludeEmptyAndAggregation(): void
     {
@@ -380,7 +380,7 @@ class TSMRANGE_Test extends PredisCommandTestCase
     /**
      * @group connected
      * @return void
-     * @requiresRedisVersion >= 8.10.0
+     * @requiresRedisVersion >= 8.9.0
      */
     public function testQueryRangeExcludesEmptySeriesWithExcludeEmptyModifierResp3(): void
     {