+++ tests/Predis/Command/Redis/TimeSeries/TSQUERYLABELS_Test.php
@@ -53,7 +53,7 @@ class TSQUERYLABELS_Test extends PredisCommandTestCase
      * @group connected
      * @group relay-resp3
      * @return void
-     * @requiresRedisVersion >= 8.10.0
+     * @requiresRedisVersion >= 8.9.0
      */
     public function testQueryReturnsLabelNamesMatchingGivenFilterExpression(): void
     {
@@ -79,7 +79,7 @@ class TSQUERYLABELS_Test extends PredisCommandTestCase
      * @group connected
      * @group relay-resp3
      * @return void
-     * @requiresRedisVersion >= 8.10.0
+     * @requiresRedisVersion >= 8.9.0
      */
     public function testQueryReturnsLabelValuesMatchingGivenFilterExpression(): void
     {
@@ -104,7 +104,7 @@ class TSQUERYLABELS_Test extends PredisCommandTestCase
     /**
      * @group connected
      * @return void
-     * @requiresRedisVersion >= 8.10.0
+     * @requiresRedisVersion >= 8.9.0
      */
     public function testQueryReturnsLabelNamesMatchingGivenFilterExpressionResp3(): void
     {
@@ -129,7 +129,7 @@ class TSQUERYLABELS_Test extends PredisCommandTestCase
     /**
      * @group connected
      * @return void
-     * @requiresRedisVersion >= 8.10.0
+     * @requiresRedisVersion >= 8.9.0
      */
     public function testQueryReturnsLabelValuesMatchingGivenFilterExpressionResp3(): void
     {