+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added support for `EXCLUDEEMPTY` argument for `TS.MRANGE` and `TS.MREVRANGE` commands
 - Added explicit testing for FT.SEARCH timeout policies
 - Added support for `TS.QUERYLABELS` command
+- Added support for `LMOVEM` and `BLMOVEM` commands
 - Added support for `FT.ALIASLIST` command
 - Added stream commands to ClusterStrategy